requirements.txt
flask
pillow
pdf2image
opencv-python
numpy
pytesseract
rapidfuzz
ultralytics
